Рефакторинг кода, упорядочил файловую структуру
