+++ src/components/task/no-task.js
@@ -1,4 +1,4 @@
-import {createElement} from "../../utils.js";
+import {createElement} from "../../utils/dom-utils.js";
 
 const createNoTaskTemplate = () => {
   return `<p class="board__no-tasks">
