+++ src/components/task/task-list.js
@@ -1,4 +1,4 @@
-import {createElement} from "../../utils.js";
+import {createElement} from "../../utils/dom-utils.js";
 
 const createTaskListTemplate = () => {
   return `<div class="board__tasks"></div>`;
