+++ src/main.js
@@ -1,7 +1,6 @@
-
 import TaskView from "./components/task/task.js";
 import TaskEditView from "./components/task/task-edit.js";
-import PageMenuView from "./components/page-menu/page-menu.js";
+import MenuView from "./components/menu/menu.js";
 import LoadMoreButtonView from "./components/load-more-button/load-more-button.js";
 import TaskBoardView from "./components/task/task-board.js";
 import TaskListView from "./components/task/task-list.js";
@@ -12,14 +11,13 @@ import FilterView from "./components/filter/filter.js";
 import {generateTask} from "./mock/task.js";
 import {generateFilters} from "./mock/filter.js";
 
-import {render, RenderPosition} from "./utils.js";
+import {render} from "./utils/dom-utils.js";
 
-const {BEFOREEND} = RenderPosition;
 const TASK_CARD_COUNT = 12;
 const TASK_COUNT_PER_STEP = 8;
 
 const tasks = new Array(TASK_CARD_COUNT).fill().map(generateTask);
-const {length: taskLength} = tasks;
+const tasksAmount = tasks.length;
 const filters = generateFilters(tasks);
 
 const pageMainElement = document.querySelector(`.main`);
@@ -52,37 +50,37 @@ const renderTask = (taskListElement, task) => {
     .querySelector(`form`)
     .addEventListener(`submit`, (evt) => {
       evt.preventDefault();
-      replaceTaskState(taskEditComponent, taskComponent);
+      replaceTaskState(taskComponent, taskEditComponent);
       document.addEventListener(`keydown`, onEscKeyDown);
     });
 
-  render(taskListElement, taskComponent.getElement(), BEFOREEND);
+  render(taskListElement, taskComponent.getElement());
 };
 
-render(pageHeaderElement, new PageMenuView().getElement(), BEFOREEND);
-render(pageMainElement, new FilterView(filters).getElement(), BEFOREEND);
+render(pageHeaderElement, new MenuView().getElement());
+render(pageMainElement, new FilterView(filters).getElement());
 
 const boardComponent = new TaskBoardView();
-render(pageMainElement, boardComponent.getElement(), BEFOREEND);
+render(pageMainElement, boardComponent.getElement());
 
 if (tasks.every((task) => task.isArchive)) {
-  render(boardComponent.getElement(), new NoTaskView().getElement(), BEFOREEND);
+  render(boardComponent.getElement(), new NoTaskView().getElement());
 } else {
-  render(boardComponent.getElement(), new SortView().getElement(), BEFOREEND);
+  render(boardComponent.getElement(), new SortView().getElement());
 
   const taskListComponent = new TaskListView();
-  render(boardComponent.getElement(), taskListComponent.getElement(), BEFOREEND);
+  render(boardComponent.getElement(), taskListComponent.getElement());
 
-  for (let i = 0; i < Math.min(taskLength, TASK_COUNT_PER_STEP); i++) {
+  for (let i = 0; i < Math.min(tasksAmount, TASK_COUNT_PER_STEP); i++) {
     renderTask(taskListComponent.getElement(), tasks[i]);
   }
 
-  if (taskLength > TASK_COUNT_PER_STEP) {
+  if (tasksAmount > TASK_COUNT_PER_STEP) {
     let renderedTaskCount = TASK_COUNT_PER_STEP;
 
     const loadMoreButtonComponent = new LoadMoreButtonView();
 
-    render(boardComponent.getElement(), loadMoreButtonComponent.getElement(), BEFOREEND);
+    render(boardComponent.getElement(), loadMoreButtonComponent.getElement());
 
     loadMoreButtonComponent.getElement()
       .addEventListener(`click`, (evt) => {
@@ -92,7 +90,7 @@ if (tasks.every((task) => task.isArchive)) {
           .forEach((task) => renderTask(taskListComponent.getElement(), task));
 
         renderedTaskCount += TASK_COUNT_PER_STEP;
-        if (renderedTaskCount >= taskLength) {
+        if (renderedTaskCount >= tasksAmount) {
           loadMoreButtonComponent.getElement().remove();
           loadMoreButtonComponent.removeElement();
         }
